pyproject.toml
[project]
name = "ecss-mcp-server"
version = "0.1.0"
description = "European Cooperation for Space Standardization (ECSS) MCP server for easy access of ECSS documents."
readme = "README.md"
requires-python = ">=3.14"
license = {file = "LICENSE"}
authors = [
    {name = "Benedict Rose"},
]
keywords = ["ecss", "mcp", "space", "standards"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: Implementation :: CPython",
    "Typing :: Typed",
]
dependencies = [
    "fastmcp>=3.2.4",
    "numpy>=2.0.0",
    "pandas>=3.0.1",
    "python-docx>=1.2.0",
    "spire-doc>=14.4.2",
]

[project.urls]
Homepage = "https://github.com/aerospace-mcp-tools/ecss-mcp-server"
Repository = "https://github.com/aerospace-mcp-tools/ecss-mcp-server"

[project.scripts]
ecss-mcp-server = "ecss_mcp_server.ecss_mcp_server:main"
ecss-parser = "ecss_mcp_server.document_parser:main"

[dependency-groups]
dev = [
    "mypy>=1.20.2",
    "pytest>=8.4.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.15.12",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: not covered",
    "@overload",
]

[tool.coverage.run]
branch = true
relative_files = true
omit = [
    ".tox/*",
    "tests/*",
]

[tool.pytest.ini_options]
minversion = "6.0"
log_cli = true
log_cli_level = "INFO"
addopts = """
--doctest-modules \
--cov=./ \
--cov-append \
--cov-report html:tests/reports/coverage-html \
--cov-report xml:tests/reports/coverage.xml \
--ignore=docs/
"""
testpaths = [
    "tests",
]

[tool.ruff]
lint.select = ["ALL"]
lint.ignore = [
    "COM812",  # Missing trailing comma (disabled due to formatter conflict)
    "D203",    # One blank line before class docstring
    "D212",    # Multi-line summary first line
    "ISC001",  # Single-line implicit string concatenation (disabled due to formatter conflict)
    "PLR0913", # Too many arguments
    "Q000",    # Single quotes found but double quotes preferred
]
line-length = 120
show-fixes = true
src = ["src"]

[tool.ruff.lint.flake8-copyright]
author = "Benedict Rose"

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
multiline-quotes = "double"

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.per-file-ignores]
# https://beta.ruff.rs/docs/rules/
"__init__.py" = ["F401", "F403", "F405", "PGH003"]
"tests/*" = ["ANN", "ARG", "INP001", "S101"]

[tool.ruff.lint.pylint]
max-args = 15
max-branches = 20
max-returns = 10
max-statements = 80

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.uv]
cache-keys = [
    { file = "pyproject.toml" },
    { git = { commit = true, tags = true } },
]

[tool.uv.build-backend]
module-name = ["ecss_mcp_server"]

[build-system]
requires = ["uv_build>=0.11.7,<0.12"]
build-backend = "uv_build"